# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Skills Latvia 2022 API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/clients": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "description": "Register a new Client",
                "operationId": "authRegister",
                "requestBody": {
                    "description": "Create a new Client",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "name",
                                    "surname"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Gibril"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Pavlovic"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succesfully logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "plainTextToken": {
                                            "type": "string",
                                            "example": "5|WhdOuIUQUwiRssfaGbdhfgGQLVF2gyBC0Z3EA7sdbm4vxFS"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/clients/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign in",
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succesfully logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "plainTextToken": {
                                            "type": "string",
                                            "example": "5|WhdOuIUQUwiRssfaGbdhfgGQLVF2gyBC0Z3EA7sdbm4vxFS"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email or Password is incorrect"
                    }
                }
            }
        },
...
        "/sorting-centers": {
            "get": {
                "tags": [
                    "SortingCenter"
                ],
                "summary": "Displays SortingCenter list",
                "description": "Displays SortingCenter list",
                "operationId": "SortingCenterList",
                "requestBody": {
                    "description": "Page number",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "page": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SortingCenterResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "SortingCenter"
                ],
                "summary": "Create a new SortingCenter",
                "description": "Create a new SortingCenter",
                "operationId": "SortingCenterStore",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SortingCenterStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SortingCenterResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sorting-centers/{id}": {
            "get": {
                "tags": [
                    "SortingCenter"
                ],
                "summary": "Displays SortingCenter by id",
                "description": "Displays SortingCenter by id",
                "operationId": "SortingCenterShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Object ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "3"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SortingCenterResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "SortingCenter"
                ],
                "summary": "Update an existing SortingCenter",
                "description": "Update an existing SortingCenter",
                "operationId": "SortingCenterUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Object ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "3"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SortingCenterUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SortingCenterResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "SortingCenter"
                ],
                "summary": "Delete SortingCenter by id",
                "description": "Delete SortingCenter by id",
                "operationId": "SortingCenterDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Object ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "3"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/NullResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NullResource": {},
...
            "SortingCenterStoreRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Kapital sorting"
                    },
                    "address": {
                        "description": "Get the validation rules that apply to the request.",
                        "type": "string",
                        "example": "Valnu iela 56"
                    }
                },
                "type": "object"
            },
            "SortingCenterUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Kapital sorting"
                    },
                    "address": {
                        "description": "Get the validation rules that apply to the request.",
                        "type": "string",
                        "example": "Valnu iela 56"
                    }
                },
                "type": "object"
            },
...
            "SortingCenterResource": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Kapital sorting"
                    },
                    "address": {
                        "type": "string",
                        "example": "Valnu iela 56"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}